chore: reorganizado DTOs do projeto
