+++ src/main/java/br/com/fatecmc/tecladostorelesbackend/presentation/controllers/ClienteController.java
@@ -3,10 +3,9 @@ package br.com.fatecmc.tecladostorelesbackend.presentation.controllers;
 import br.com.fatecmc.tecladostorelesbackend.domain.services.ClienteService;
 import br.com.fatecmc.tecladostorelesbackend.presentation.dtos.input.ClienteCadastroDTO;
 import br.com.fatecmc.tecladostorelesbackend.presentation.dtos.input.ClienteEditadoDTO;
-import br.com.fatecmc.tecladostorelesbackend.presentation.dtos.output.CartaoDTO;
-import br.com.fatecmc.tecladostorelesbackend.presentation.dtos.output.ClienteDTO;
+import br.com.fatecmc.tecladostorelesbackend.presentation.dtos.input.CartaoDTO;
 import br.com.fatecmc.tecladostorelesbackend.presentation.dtos.output.ClienteRetornoDTO;
-import br.com.fatecmc.tecladostorelesbackend.presentation.dtos.output.EnderecoDTO;
+import br.com.fatecmc.tecladostorelesbackend.presentation.dtos.input.EnderecoDTO;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -47,11 +46,6 @@ public class ClienteController {
     @PutMapping("/{id}")
     public ClienteRetornoDTO update(@PathVariable("id") Long id, @RequestBody ClienteEditadoDTO cliente){
         return this.clienteService.update(id, cliente);
-    }
-
-    @PatchMapping("/{id}/cartao")
-    public ClienteDTO patchWithCartao(@PathVariable("id") Long id, @RequestBody CartaoDTO cartao){
-        return this.clienteService.patchWithCartao(id, cartao);
     }
 
     @PatchMapping("/{id}/enderecos/{idEndereco}")
