--- src/main/java/br/com/fatecmc/tecladostorelesbackend/presentation/dtos/input/EnderecoInputDTO.java
@@ -1,4 +1,0 @@
-package br.com.fatecmc.tecladostorelesbackend.presentation.dtos.input;
-
-public class EnderecoInputDTO {
-}
